more profile refactoring
